requirements.txt
requests~=2.32.3
selenium~=4.14.0
chromedriver_autoinstaller
pytdm
pyttsx3~=2.90
edge-tts
pydub
opencv-python
Pillow~=10.0.1
ffpyplayer
scipy
moviepy~=1.0.3
numpy~=1.26.1
beautifulsoup4~=4.12.3